requirements.txt
# Core dependencies for Pi-hole LLM Analytics
requests>=2.31.0
python-dotenv>=1.0.0

# Data analysis (optional for enhanced analytics)
# pandas>=2.0.0
# numpy>=1.24.0

# Type checking support
typing-extensions>=4.7.0

# Development dependencies (for contributors)
# pytest>=7.4.0
# pylint>=2.17.0
# black>=23.0.0
# mypy>=1.5.0
